+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add `concat` function
 * Add `Signal#concat` method
 * Add `always` function
+* Add `cycle` function
 
 ## 1.4.0 (2019-01-08)